Open Nectar 0.6: port a Android, packs de texturas HD y renderizador residente
Android (docs/ANDROID_PLAN.md, fases 0-8): proyecto Gradle en android/ que
compila el arbol desde el CMakeLists de la raiz con SDL2 vendido
(third_party/SDL2-android), capa pc_port/android (logcat, rutas, ciclo de
vida, instalador JNI sobre pikmin_disc_image), un APK con USA y PAL, firma de
release fuera del repositorio y packaging/android/package-apk.sh.

Controles tactiles (pc_port/touch, docs/CONTROLES_TACTILES_PLAN.md): stick
flotante, botones con editor de disposicion, gestos de camara, menus tocando
lo que dibuja el juego, e iconos tacticos en lugar de los botones de GameCube
en cualquier pantalla 2D (P2DPicture sustituye las texturas de boton y la
capa muestra ese boton mientras el aviso esta en pantalla). Atras en el mapa
vuelve al selector de partida.

Rendimiento (docs/PLAN_RENDIMIENTO.md): mallas residentes en GPU con skinning
en el vertex shader, anillo de VBO con fences en GLES, cache de binarios de
shader, invalidacion de framebuffer, heaps ovl/app sin tocar paginas.

Packs de texturas (docs/PLAN_TEXTURAS_HD.md): cargador DDS/PNG con formato
de nombres de Dolphin (pc_port/gl/pc_texpack), bcdec para BC1-7 sin soporte
del driver, menu F1 e instalacion desde ZIP/RAR en Android.

+++ src/jaudio/aramcall.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include "jaudio/aramcall.h"
 
 #include "jaudio/dummyrom.h"
@@ -99,7 +100,7 @@ void Init_AramMotherHeap(void)
 		inited = TRUE;
 
 		void* alloc = ARAllocFull(&outSize);
-		Jac_InitMotherHeap(&aram_mother, (u32)alloc, outSize, 0);
+		Jac_InitMotherHeap(&aram_mother, (u32)(uintptr_t)alloc, outSize, 0);
 	}
 }
